refactor, moved methods around to more appropriate classes, added comments, removed unnecessary lines
