+++ src/main/java/com/kpme/kpme/InputPattern.java
@@ -1,11 +1,15 @@
 package com.kpme.kpme;
 
+
+// class to hold inputPattern
 public class InputPattern {
     private String inputPattern;
 
     public void setInputPattern(){
         this.inputPattern = inputPattern;
     }
+
+    // getter
     public String getInputPattern() {
         return inputPattern;
