+++ src/main/java/com/kpme/kpme/KpmeApplication.java
@@ -3,11 +3,6 @@ package com.kpme.kpme;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-import java.util.Arrays;
-
-import static com.kpme.kpme.Row.generateRow;
-import static com.kpme.kpme.Row.parsePattern;
-
 @SpringBootApplication
 public class KpmeApplication {
