+++ src/main/java/com/kpme/kpme/PatternHandlerController.java
@@ -1,16 +1,11 @@
 package com.kpme.kpme;
 
-
 import org.springframework.web.bind.annotation.*;
-
-import java.lang.reflect.Array;
 import java.util.ArrayList;
 
 
 @RestController
 public class PatternHandlerController {
-
-//    private static final String template = "Here is your pattern: %s!";
 
     @CrossOrigin(origins = "*")
     @PostMapping("")
