+++ src/main/java/com/kpme/kpme/Stitch.java
@@ -1,5 +1,7 @@
 package com.kpme.kpme;
 
+
+// class to hold a single stitch
 public class Stitch {
     public String stitch;
